requirements.txt
tensorflow
keras
numpy
opencv-python
streamlit
Pillow
pandas
matplotlib